+++ ent/platformaccount/where.go
@@ -8,47 +8,47 @@ import (
 )
 
 // ID filters vertices based on their ID field.
-func ID(id int) predicate.PlatformAccount {
+func ID(id uint64) predicate.PlatformAccount {
 	return predicate.PlatformAccount(sql.FieldEQ(FieldID, id))
 }
 
 // IDEQ applies the EQ predicate on the ID field.
-func IDEQ(id int) predicate.PlatformAccount {
+func IDEQ(id uint64) predicate.PlatformAccount {
 	return predicate.PlatformAccount(sql.FieldEQ(FieldID, id))
 }
 
 // IDNEQ applies the NEQ predicate on the ID field.
-func IDNEQ(id int) predicate.PlatformAccount {
+func IDNEQ(id uint64) predicate.PlatformAccount {
 	return predicate.PlatformAccount(sql.FieldNEQ(FieldID, id))
 }
 
 // IDIn applies the In predicate on the ID field.
-func IDIn(ids ...int) predicate.PlatformAccount {
+func IDIn(ids ...uint64) predicate.PlatformAccount {
 	return predicate.PlatformAccount(sql.FieldIn(FieldID, ids...))
 }
 
 // IDNotIn applies the NotIn predicate on the ID field.
-func IDNotIn(ids ...int) predicate.PlatformAccount {
+func IDNotIn(ids ...uint64) predicate.PlatformAccount {
 	return predicate.PlatformAccount(sql.FieldNotIn(FieldID, ids...))
 }
 
 // IDGT applies the GT predicate on the ID field.
-func IDGT(id int) predicate.PlatformAccount {
+func IDGT(id uint64) predicate.PlatformAccount {
 	return predicate.PlatformAccount(sql.FieldGT(FieldID, id))
 }
 
 // IDGTE applies the GTE predicate on the ID field.
-func IDGTE(id int) predicate.PlatformAccount {
+func IDGTE(id uint64) predicate.PlatformAccount {
 	return predicate.PlatformAccount(sql.FieldGTE(FieldID, id))
 }
 
 // IDLT applies the LT predicate on the ID field.
-func IDLT(id int) predicate.PlatformAccount {
+func IDLT(id uint64) predicate.PlatformAccount {
 	return predicate.PlatformAccount(sql.FieldLT(FieldID, id))
 }
 
 // IDLTE applies the LTE predicate on the ID field.
-func IDLTE(id int) predicate.PlatformAccount {
+func IDLTE(id uint64) predicate.PlatformAccount {
 	return predicate.PlatformAccount(sql.FieldLTE(FieldID, id))
 }
